feat: reorganize DEX api
